ASoC: sof_es8336: reduce pop noise on speaker
The Speaker need to be turned on slightly behind the codec turned on.
It also need to be turned off slightly before the codec turned down.
Current code uses delay in DAPM_EVENT to do it but the mdelay delays the
DAPM itself and thus has no effect. A delayed_work is added to turn on the
speaker.
The speaker is turned off in .trigger since trigger is called slightly before
the DAPM events.

Signed-off-by: Zhu Ning <zhuning@everest-semi.com>