fixup! ASoC: SoundWire codecs: make regmap cache-only in probe
There are two regmap in rt5682_priv struct, one is sdw_regmap which is
for IO transfer, and the other is used for registers control.
We need to set both regmaps when we set cache only. Because if we set
rt5682->sdw_regmap only, rt5682->regmap won't set/get the right value
when it call regmap_write/read(rt5682->sdw_regmap, ...). If we set
rt5682->regmap only, regmap_write(rt5682->sdw_regmap, ...) is used
in rt5682_clock_config which will be called by the ..bus_config ops.

Signed-off-by: Bard Liao <yung-chuan.liao@linux.intel.com>